requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.4
urllib3==2.1.0
openpyxl==3.1.2
ldap3>=2.9.1
PyJWT>=2.8.0